[WIP] Start work on a memory mapped file storage backend
The current implementation will do all read/writes directly to disc when
managing the database. The main idea with this implementation is to use
MappedByteBuffer which will load a file into memory that can be accessed
faster. The current implementation status is that it is functionally
finished but not usable due to bugs.

Fixes #62
Fixes #63
Fixes #66